app/requirements.txt
feedparser==6.0.11
httpx==0.27.2
trafilatura==1.8.0
PyYAML==6.0.2
summa==1.2.0
tldextract==5.1.2
